refactor: altered people hooks
