+++ src/lib/hooks/people/usePeople.ts
@@ -8,12 +8,3 @@ export function usePeople() {
     staleTime: Infinity,
   })
 }
-
-export function usePeopleById(id: string) {
-  return useQuery({
-    queryKey: ['people', id],
-    queryFn: () => peoplesService.getById(id),
-    staleTime: Infinity,
-    enabled: !!id,
-  })
-}
