+++ src/lib/hooks/people/usePeopleById.ts
@@ -1,0 +1,11 @@
+import { peoplesService } from '@lib/services/peoples'
+import { useQuery } from '@tanstack/react-query'
+
+export function usePeopleById(id: string) {
+  return useQuery({
+    queryKey: ['people', id],
+    queryFn: () => peoplesService.getById(id),
+    staleTime: Infinity,
+    enabled: !!id,
+  })
+}
